+++ packages/crypto_wallet_util/lib/transaction.dart
@@ -56,7 +56,14 @@ export 'src/transaction/xrp/xrp_lib.dart';
 export 'src/transaction/filecoin/fil_lib.dart';
 export 'src/transaction/sc/sc_lib.dart';
 export 'src/transaction/sc/sc_wasm_bridge.dart';
+export 'src/transaction/sc/sc_go_ffi_bridge.dart';
 export 'src/transaction/sc/sc_wasm_run_bridge.dart';
+
+/// export scp module
+export 'src/transaction/scp/tx_signer.dart';
+export 'src/transaction/scp/tx_data.dart';
+export 'src/transaction/scp/scp_lib.dart';
+export 'src/transaction/scp/scp_sig_hash.dart';
 
 /// export btc psbt
 export 'src/transaction/btc/psbt.dart';